feat: move market header stats comp to appropriate location, add spot header
