+++ apps/trading/components/market-header/market-header-spot.tsx
@@ -1,0 +1,3 @@
+export const MarketHeaderSpot = () => {
+  return <div>spot</div>;
+};
