+++ apps/trading/components/market-header/market-header-stats.tsx
