+++ apps/trading/components/market-header/market-header.tsx
@@ -2,10 +2,16 @@ import { Popover, VegaIcon, VegaIconNames } from '@vegaprotocol/ui-toolkit';
 import { Header, HeaderTitle } from '../header';
 import { Route, Routes, useParams } from 'react-router-dom';
 import { MarketSelector } from '../../components/market-selector/market-selector';
-import { MarketHeaderStats } from '../../client-pages/market/market-header-stats';
-import { useMarket, useMarketList } from '@vegaprotocol/markets';
+import { MarketHeaderStats } from './market-header-stats';
+import {
+  type Market,
+  getProductType,
+  useMarket,
+  useMarketList,
+} from '@vegaprotocol/markets';
 import { useState } from 'react';
 import { ProductTypeShortName } from '@vegaprotocol/types';
+import { MarketHeaderSpot } from './market-header-spot';
 
 export const MarketHeader = () => {
   const { marketId } = useParams();
@@ -53,10 +59,28 @@ export const MarketHeader = () => {
               </Popover>
             }
           >
-            <MarketHeaderStats market={data} />
+            <MarketHeaderSwitch market={data} />
           </Header>
         }
       />
     </Routes>
   );
+};
+
+export const MarketHeaderSwitch = ({ market }: { market: Market }) => {
+  const productType = getProductType(market);
+
+  if (productType === 'Spot') {
+    <MarketHeaderSpot />;
+  }
+
+  if (productType === 'Future') {
+    return <MarketHeaderStats market={market} />;
+  }
+
+  if (productType === 'Perpetual') {
+    return <MarketHeaderStats market={market} />;
+  }
+
+  throw new Error('invalid product type');
 };
